requirements.txt
flask==3.0.3
mysql-connector-python==8.4.0
flask_cors==4.0.1
